Add TLD > 1 extraction. Applied patch from Sebdraven.
This is still work in progress. Some broken stuff need to be fixed before the release.

+++ src/lib/bindings/python/pyfaup/functions.py
@@ -2,11 +2,11 @@
 # Copyright (C) Sebastien Tricaud 2012
 
 from ctypes import *
-from bind import *
+from .bind import *
 
 faup_get_version = bind("faup_get_version", c_char_p)
 faup_init = bind("faup_init", c_void_p)
-faup_decode = bind("faup_decode", None, c_void_p, c_char_p, c_uint)
+faup_decode = bind("faup_decode", None, c_void_p, c_char_p, c_size_t)
 faup_terminate = bind("faup_terminate", None, c_void_p)
 
 faup_get_scheme_pos = bind("faup_get_scheme_pos", c_int, c_void_p)
